added find path method
